feat: allocate multiple streams per endpoint
Extract Layers data structures into a separate file.

Creates SingleSourceAllocation2.kt which is to be used
by the multi-stream support flow. Eventually this will
be merged later once all changes are identified and
should we still agree it makes sense to merge (as opposed
to keeping separate and deleting old classes once deprecated).

Branches off the multi stream feature flag for the allocation
algorithm in the BandwidthAllocator.java.